+++ time/src/types/CreateTimeRecord.ts
@@ -1,0 +1,7 @@
+export type CreateTimeRecord = {
+  projectId: number;
+  date: string;
+  description: string;
+  duration: number;
+  onSite: boolean;
+};