+++ time/CHANGELOG.md
@@ -1,3 +1,5 @@
 # time Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [TimeRecords] - 2025-04-24
+
+- Added TimeRecords read, create and delte